// src/app/types/ResTypes.ts
export interface VerifyOtpRes {
  status: boolean;
  data: {
    token: string;
    user: {
      user_name: string;
      email: string;
      mobile_num: string;
      profile_pic: string;
      id_proof: string;
      selfie: string;
      dob: string;
      intrests: string[];
      platforms: string[];
      user_id: number;
      full_name: string;
      first_name: string;
      last_name: string;
      country_code: string;
      socket_id: string;
      otp: number;
      password: string;
      login_type: string;
      gender: string;
      country: string;
      country_short_name: string;
      state: string;
      city: string;
      bio: string;
      device_token: string;
      profile_verification_status: boolean;
      login_verification_status: boolean;
      is_private: boolean;
      is_admin: boolean;
      available_coins: string;
      account_name: string;
      account_number: string;
      bank_name: string;
      swift_code: string;
      IFSC_code: string;
      total_socials: number;
      blocked_by_admin: boolean;
      createdAt: string;
      updatedAt: string;
    };
  };
  message: string;
  toast: boolean;
}